Complete initial checks for submissions.

There are several initial checks undertaken for each submission. These checks reflect the requests made in the author submission guide. Results will be show in the initial_checks.log file in the submission folder

initial_checks(dic = "en_US")

initial_check_article(path = here::here(), dic = "en_US")

check_wrappers(path)

check_filenames(path)

check_unnecessary_files(path)

check_cover_letter(path)

check_title(path)

check_section(path)

check_abstract_before_intro(path)

check_spelling(path, dic = "en_US")

check_proposed_pkg()

check_packages_available(path)

output_summary(path, file = stdout())

Arguments

dic
the dictionary used for spelling check. See dict argument in hunspell::hunspell()

path
The directory that contains the .tex file (Ideally, this directory should contain .bib, .rmd, and .tex with author names and two RJwrapper files: RJwrapper.pdf and RJwrapper.tex)

Details

Folder structure checks:

check_wrappers(): Check that the two expected RJwrapper files: RJwrapper.tex and RJwrapper.pdf exist

check_filenames(): Check that the .bib, .rmd, and .tex all presents and have consistent names

check_unnecessary_files(): Check that RJtemplate.tex and RJournal.sty are not included in the directory

Tex file checks:

check_title(): article title is in title case

check_section(): section sections are in sentence case

check_abstract_before_intro(): abstract comes before the introduction section

check_spelling(): potential spelling mistakes

check_proposed_pkg(): package proposed in the paper is on CRAN

check_packages_available(): Check that packages mentioned in the text are available on CRAN