R Markdown output formats for R Journal articles

The R Journal is built upon the distill framework with some modifications. This output format behaves almost identically to the distill::distill_article() format, with some formatting and structural changes.

rjournal_web_article(
  toc = FALSE,
  toc_depth = 3,
  toc_float = TRUE,
  fig_width = 6.5,
  fig_height = 4,
  fig_retina = 2,
  fig_caption = TRUE,
  dev = "png",
  smart = TRUE,
  code_folding = FALSE,
  self_contained = TRUE,
  highlight = "default",
  highlight_downlit = TRUE,
  mathjax = "default",
  extra_dependencies = NULL,
  theme = NULL,
  css = NULL,
  includes = NULL,
  keep_md = FALSE,
  lib_dir = NULL,
  md_extensions = NULL,
  pandoc_args = NULL,
  ...
)

rjournal_pdf_article(...)

Arguments

toc
TRUE to include a table of contents in the output

toc_depth
Depth of headers to include in table of contents

toc_float
Float the table of contents to the left when the article is displayed at widths > 1000px. If set to FALSE or the width is less than 1000px the table of contents will be placed above the article body.

fig_width
Default width (in inches) for figures

fig_height
Default height (in inches) for figures

fig_retina
Scaling to perform for retina displays (defaults to 2, which currently works for all widely used retina displays). Set to NULL to prevent retina scaling. Note that this will always be NULL when keep_md is specified (this is because fig_retina relies on outputting HTML directly into the markdown document).

fig_caption
TRUE to render figures with captions

dev
Graphics device to use for figure output (defaults to png)

smart
Produce typographically correct output, converting straight quotes to curly quotes, --- to em-dashes, -- to en-dashes, and ... to ellipses.

code_folding
Include code blocks hidden, and allow users to optionally display the code by clicking a "Show code" button just above the output. Pass a character vector to customize the text of the "Show code" button. You can also specify code_folding as chunk option for per-chunk folding behavior.

self_contained
Produce a standalone HTML file with no external dependencies, using data: URIs to incorporate the contents of linked scripts, stylesheets, images, and videos. Note that even for self contained documents MathJax is still loaded externally (this is necessary because of its size).

highlight
Syntax highlighting style. Supported styles include "default", "rstudio", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", "breezedark", and "haddock". Pass NULL to prevent syntax highlighting.

highlight_downlit
Use the downlit package to highlight R code (including providing hyperlinks to function documentation).

mathjax
Include mathjax. The "default" option uses an https URL from a MathJax CDN. The "local" option uses a local version of MathJax (which is copied into the output directory). You can pass an alternate URL or pass NULL to exclude MathJax entirely.

extra_dependencies
Additional function arguments to pass to the base R Markdown HTML output formatter html_document_base

theme
CSS file with theme variable definitions

css
CSS and/or Sass files to include. Files with an extension of .sass or .scss are compiled to CSS via sass::sass().

includes
Named list of additional content to include within the document (typically created using the includes function).

keep_md
Keep the markdown file generated by knitting.

lib_dir
Directory to copy dependent HTML libraries (e.g. jquery, bootstrap, etc.) into. By default this will be the name of the document with _files appended to it.

md_extensions
Markdown extensions to be added or removed from the default definition or R Markdown. See the rmarkdown_format for additional details.

pandoc_args
Additional command line options to pass to pandoc

...
Arguments passed to distill::distill_article() for web articles, and rticles::rjournal_article() for pdf articles.