A single article check

A single article check

initial_check_article(
  path,
  dic = "en_US",
  pkg,
  ...,
  ask = interactive(),
  logfile = file.path(path, "initial_checks.log")
)

check_filenames(path)

check_structure(path)

check_unnecessary_files(path)

check_cover_letter(path)

check_title(path, ignore = "")

check_section(path)

check_abstract_before_intro(path)

check_spelling(path, dic = "en_US", ...)

check_proposed_pkg(pkg, ask = interactive())

check_packages_available(path, ignore)

Arguments

path

string, path to the directory that contains the .tex file (Ideally, this directory should contain .bib, .Rmd, and .tex with author names and two RJwrapper files: RJwrapper.pdf and RJwrapper.tex)

dic

string, the dictionary used for spelling check. See dict argument in hunspell::hunspell()

pkg

string, optional. The name of the proposed package (if relevant), to be checked for activity on CRAN

...

additional arguments for spelling check with hunspell::hunspell

ask

logical, if TRUE then checks may ask the user for interactive input of missing information.

logfile

a connection for the output log, or a string with the filename of the output log or NULL if no log should be written

ignore

The words to ignore in title check, use c(pkg, pkg, ...) for multiple quoted words

Value

list of all results (see log_error for details). You can use unlist() to get a character vector of the result statuses.

Details

Folder structure checks:

check_filenames(): the three files (.bib, .Rmd, and .tex) all present and have consistent names

check_structure(): check validity of all filenames and depth of the directory structure

check_unnecessary_files(): the template file (i.e., RJtemplate.tex) is not included in the directory

check_cover_letter(): a motivational letter

Content checks:

check_title(): article title is in title case

check_section(): section sections are in sentence case

check_abstract_before_intro(): abstract comes before the introduction section

check_spelling(): potential spelling mistakes

check_proposed_pkg(): package proposed in the paper is on CRAN

check_packages_available(): packages mentioned in the article are available on CRAN

See vignette("create_article", package = "rjtools") for how to use the check functions

Examples

your_article_path <- system.file("sample-article", package = "rjtools")
if (interactive()) initial_check_article(your_article_path)